+++ CHANGELOG.md
@@ -1,1 +1,17 @@
 
+1.1.0
+-
+Add Parcool and Project MMO Integration. Levels 0 through 12  in acrobatics unlock different parcool abilities.
+- Map.entry(0L, List.of(RideZipline.class, SkyDive.class, QuickTurn.class)),
+- Map.entry(1L, List.of(Crawl.class)),
+- Map.entry(2L, List.of(ChargeJump.class)),
+- Map.entry(3L, List.of(FastRun.class, Slide.class)),
+- Map.entry(4L, List.of(Vault.class)),
+- Map.entry(5L, List.of(WallJump.class)),
+- Map.entry(6L, List.of(ClingToCliff.class, ClimbPoles.class, ClimbUp.class, JumpFromBar.class, HangDown.class)),
+- Map.entry(7L, List.of(BreakfallReady.class, Tap.class, Roll.class)),
+- Map.entry(8L, List.of(CatLeap.class, Dodge.class)),
+- Map.entry(9L, List.of(WallSlide.class)),
+- Map.entry(10L, List.of(Flipping.class, HideInBlock.class)),
+- Map.entry(11L, List.of(VerticalWallRun.class)),
+- Map.entry(12L, List.of(HorizontalWallRun.class))